+++ NEWS.md
@@ -5,7 +5,7 @@
 - The internal representation of index objects has changed so that both
 the `index` and `contrib` components of an index are now matrices instead of
 lists. This simplifies the code base and in some cases improves performance, but
-will break any coding that directly uses the internal components of an index.
+will break any code that directly uses the internal components of an index.
 
 - A new internal optimization means that index objects without contributions now
 consume about half as much memory.
@@ -30,6 +30,9 @@ in `elementary_index()`.
 
 - Added a new function `combine_classifications()` to more easily build an
 aggregation structure that spans multiple hierarchical classifications.
+
+- Subscripting an index object with a matrix is now more flexible and can
+return/replace a list of index objects.
 
 ## Bug fixes